# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "LTI Canvas Devops Backend",
    "description": "LTI Canvas Devops API Documentation"
  },
  "schemes": [
    "http",
    "https"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
...
  "paths": {
...
    "/pipeline/onboarding_create/save_initial_pipeline_data": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Saves initial pipeline data",
        "description": "Stores pipeline info to pipelines table and links it with user and application",
        "parameters": [
          {
            "name": "pipeline_details",
            "in": "body",
            "description": "Pipeline Detail Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "type": "string",
                  "example": "Darshan Bhojani"
                },
                "pipeline_name": {
                  "type": "string",
                  "example": "JAVA Darshan Pipe"
                },
                "application_key": {
                  "type": "string",
                  "example": "RAHULS661agm"
                },
                "pipeline_description": {
                  "type": "string",
                  "example": "JAVA Pipe"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tool Validation is successfull",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "PIP5F9SLD2"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Failed to save PipelineInfo",
            "schema": {
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Failed to save PipelineInfo for pipeline_name"
                }
              }
            }
          }
        }
      }
    },
...
    "/pipeline/onboarding_create/onboard_plan": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Onboards pipeline to Planning Tool",
        "description": "Creates a project in Planning Tool and updates the details in Pipeline table.",
        "parameters": [
          {
            "name": "pipeline_details",
            "in": "body",
            "description": "Pipeline Detail Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "instance_name": {
                  "type": "string",
                  "example": "Public Jira"
                },
                "pipeline_key": {
                  "type": "string",
                  "example": "PIP2WXRLXM"
                },
                "application_key": {
                  "type": "string",
                  "example": "RAHULS661agm"
                },
                "tool_name": {
                  "type": "string",
                  "example": "Jira"
                },
                "pipeline_name": {
                  "type": "string",
                  "example": "new test 1"
                },
                "project_name": {
                  "type": "string",
                  "example": "RAHULS661agm_new test 1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project created successfully in Jira",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Project created in Jira"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Error while creating project in tool_name",
            "schema": {
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Failed to create project in Jira"
                }
              }
            }
          }
        }
      }
    },
    "/pipeline/onboarding_create/onboard_code_quality": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Onboards pipeline to Code Quality Tool",
        "description": "Creates a project in Code Quality Tool and generates sonar-properties file and updates the details in Pipeline table.",
        "parameters": [
          {
            "name": "pipeline_details",
            "in": "body",
            "description": "Pipeline Detail Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "instance_name": {
                  "type": "string",
                  "example": "public sonar"
                },
                "pipeline_key": {
                  "type": "string",
                  "example": "PIP2WXRLXM"
                },
                "application_key": {
                  "type": "string",
                  "example": "RAHULS661agm"
                },
                "tool_name": {
                  "type": "string",
                  "example": "Sonarqube"
                },
                "pipeline_name": {
                  "type": "string",
                  "example": "new test 1"
                },
                "project_name": {
                  "type": "string",
                  "example": "RAHULS661agm_new test 1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project created successfully in Sonarqube",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Project created in Sonarqube"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Error while creating project in tool_name",
            "schema": {
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Failed to create project in Sonarqube"
                }
              }
            }
          }
        }
      }
    },
    "/pipeline/onboarding_create/onboard_continuous_integration": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Onboards pipeline to Continuous Integration Tool",
        "description": "Creates a job in Continuous Integration Tool and generates Jenkinsfile and updates the details in Pipeline table.",
        "parameters": [
          {
            "name": "pipeline_details",
            "in": "body",
            "description": "Pipeline Detail Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "instance_name": {
                  "type": "string",
                  "example": "Public Jenkins"
                },
                "pipeline_key": {
                  "type": "string",
                  "example": "PIP2WXRLXM"
                },
                "application_key": {
                  "type": "string",
                  "example": "RAHULS661agm"
                },
                "tool_name": {
                  "type": "string",
                  "example": "Jenkins"
                },
                "pipeline_name": {
                  "type": "string",
                  "example": "new test 1"
                },
                "project_name": {
                  "type": "string",
                  "example": "RAHULS661agm_new test 1"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project created successfully in Jenkins",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Project created in Jenkins"
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Error while creating project in tool_name",
            "schema": {
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Failed to create project in Jenkins"
                }
              }
            }
          }
        }
      }
    },
    "/pipeline/template_screen/get_all_templates": {
      "get": {
        "tags": [
          "Templates"
        ],
        "summary": "Get all Templates",
        "description": "Returns templates based on application name",
        "parameters": [
          {
            "name": "application_name",
            "in": "query",
            "description": "Application Name of the application",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of Templates",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Template_data"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Input Parameter"
          },
          "401": {
            "description": "The user is unauthorized to make this request"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Error while fetching templates",
            "schema": {
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Templates for application_name do not exist"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Template_data": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "5f43a0961c9cb214e889bbce"
        },
        "is_custom": {
          "type": "boolean",
          "example": true
        },
        "template_name": {
          "type": "string",
          "example": "Template Axis 2"
        },
        "application_name": {
          "type": "string",
          "example": "Application to handle Citi Banks transaction"
        },
        "application_key": {
          "type": "string",
          "example": "AT RISK"
        },
        "created_At": {
          "type": "string",
          "example": "24-08-2020",
          "description": "Number of pipelines in the application"
        },
        "plan": {
          "type": "object",
          "properties": {
            "tool_name": {
              "type": "string",
              "example": "Jira"
            },
            "instance_name": {
              "type": "number",
              "example": "Public Jira"
            },
            "instance_id": {
              "type": "string",
              "example": "5e05d2a027673c020e9bdc73"
            },
            "tool_url": {
              "type": "number",
              "example": "http://3.87.19.19:8081"
            }
          }
        },
        "scm": {
          "type": "object",
          "properties": {
            "tool_name": {
              "type": "string",
              "example": "Bitbucket"
            },
            "instance_name": {
              "type": "number",
              "example": "Public Bitbucket"
            },
            "instance_id": {
              "type": "string",
              "example": "5dd627047fa55a5924396a46"
            },
            "tool_url": {
              "type": "number",
              "example": "http://3.87.19.19:7990"
            }
          }
        },
        "code_quality": {
          "type": "object",
          "properties": {
            "tool_name": {
              "type": "string",
              "example": "Sonarqube"
            },
            "instance_name": {
              "type": "number",
              "example": "Public Sonarqube"
            },
            "instance_id": {
              "type": "string",
              "example": "5dd628167fa55a5924396b24"
            },
            "tool_url": {
              "type": "number",
              "example": "http://3.230.103.53:9000"
            }
          }
        },
        "continuous_integration": {
          "type": "object",
          "properties": {
            "tool_name": {
              "type": "string",
              "example": "Jenkins"
            },
            "instance_name": {
              "type": "number",
              "example": "Public Jenkins"
            },
            "instance_id": {
              "type": "string",
              "example": "5f2427776c4546347c213a8b"
            },
            "tool_url": {
              "type": "number",
              "example": "http://3.81.19.57:8080"
            }
          }
        }
      }
    }
  }
}